requirements.txt
torch==2.0.1
numpy==1.23.5
python-chess==1.999
tensorboard==2.13.0
pygame==2.5.0
cairosvg==2.7.1
